use long for elapsed
git-svn-id: https://svn.exedio.com/cope/persistence/trunk@19543 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ runtime/testsrc/com/exedio/cope/pattern/DispatcherItem.java
@@ -41,8 +41,8 @@ public class DispatcherItem extends Item implements Dispatchable
 	{
 		boolean fail;
 		int dispatchCount = 0;
-		int dispatchLastSuccessElapsed = 0;
-		final ArrayList<Integer> dispatchFailureElapsed = new ArrayList<Integer>();
+		long dispatchLastSuccessElapsed = 0;
+		final ArrayList<Long> dispatchFailureElapsed = new ArrayList<Long>();
 		int notifyFinalFailureCount = 0;
 
 		Log(final boolean fail)
@@ -65,10 +65,10 @@ public class DispatcherItem extends Item implements Dispatchable
 		Thread.sleep(5);
 		if(log.fail)
 		{
-			log.dispatchFailureElapsed.add((int)toMillies(nanoTime(), start));
+			log.dispatchFailureElapsed.add(toMillies(nanoTime(), start));
 			throw new IOException(getBody());
 		}
-		log.dispatchLastSuccessElapsed = ((int)toMillies(nanoTime(), start));
+		log.dispatchLastSuccessElapsed = toMillies(nanoTime(), start);
 	}
 
 	public void notifyFinalFailure(final Dispatcher dispatcher, final Exception cause)
